use int64_t to represent time. see http://rb.corp.taobao.com/r/94849/
git-svn-id: http://code.taobao.org/svn/tfs@2974 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/stable-2.5/src/dataserver/integrity_manager.cpp
@@ -113,8 +113,8 @@ namespace tfs
                 }
                 //TODO: assert(EXIT_CHECK_CRC_ERROR != ret);
                 TBSYS_LOG_DW(ret, "check block %"PRI64_PREFIX"u integrity, ret: %d", *it, ret);
-                int32_t per_block_cost =
-                  ds_info.check_integrity_interval_days_ * 86400 / info->total_main_block_count_;
+                int64_t per_block_cost =
+                  static_cast<int64_t>(ds_info.check_integrity_interval_days_) * 86400 / info->total_main_block_count_;
                 interruptable_usleep(per_block_cost * 1000000);
               }
               else
